+++ anki_addon/__init__.py
@@ -2,7 +2,7 @@ from aqt import mw, gui_hooks
 from aqt.utils import qconnect
 from aqt.qt import QAction
 
-from .ui import SonaveebDialog
+from .ui import SonaveebDialog, check_templates_updates
 from .sonaveeb import Sonaveeb
 
 def open_sonaveeb_dialog():
@@ -23,3 +23,5 @@ qconnect(action.triggered, open_sonaveeb_dialog)
 mw.form.menuTools.addAction(action)
 
 gui_hooks.profile_will_close.append(destroy_sonaveeb_dialog)
+gui_hooks.profile_did_open.append(check_templates_updates)
+