+++ anki_addon/ui/__init__.py
@@ -1,1 +1,2 @@
 from .main_window import SonaveebDialog
+from .misc import check_templates_updates